third_party/blink/web_tests/fast/canvas/font-size-precision.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>
canvas {
  font-family: sans-serif;
  font-size: 100px;
  zoom: 3;
}
</style>
<canvas width="100" height="100"></canvas>
<script>
test(function() {
  var ctx = document.querySelector('canvas').getContext('2d');
  // Set same font as will be computed on the canvas element.
  ctx.font = "10.5px monospace";
  assert_equals(ctx.font, "10.5px monospace");
  ctx.font2 = "10px monospace";
  assert_equals(ctx.font2, "10px monospace");
}, "The precison of the font should remain the same.");
</script>
